Safari-611.1.10/LayoutTests/inspector/css/node-styles-refreshed-expected.txt

Testing that WI.DOMNodeStyles.Event.Refreshed event dispatches with correct significantChange flag.


== Running test suite: NodeStylesRefreshed
-- Running test case: NodeStylesRefreshed.ClassAdded
PASS: Adding '.baz' class should cause a significant change.

-- Running test case: NodeStylesRefreshed.ClassRemoved
PASS: Removing '.foo' class should cause a significant change.

-- Running test case: NodeStylesRefreshed.IrrelevantClassAdded
PASS: Adding '.blah' class shouldn't cause a significant change.

-- Running test case: NodeStylesRefreshed.IrrelevantClassRemoved
PASS: Removing '.blah' class shouldn't cause a significant change.
